+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Added
+
+- Account settings page with profile, notification, and security placeholders
+
+### Changed
+
+- Account menu now routes to the settings page
+
 ## [0.1.0] - 2026-02-04
 
 ### Added